always return error when wrong options are used.

+++ dos2unix/dos2unix.c
@@ -817,8 +817,8 @@ int ConvertDosToUnixNewFile(char *ipInFN, char *ipOutFN, CFlag *ipFlag)
 #endif
     if (!ipFlag->Quiet)
     {
+      ipFlag->error = errno;
       perror(_("dos2unix: Failed to open temporary output file"));
-      ipFlag->error = 1;
     }
     RetVal = -1;
   }
@@ -888,8 +888,8 @@ int ConvertDosToUnixNewFile(char *ipInFN, char *ipOutFN, CFlag *ipFlag)
     }
   }
 
-  /* any error? */
-  if (RetVal)
+  /* any error? cleanup the temp file */
+  if (RetVal && (TempPath != NULL))
   {
     if (unlink(TempPath))
     {
@@ -1029,8 +1029,8 @@ int ConvertDosToUnixOldFile(char* ipInFN, CFlag *ipFlag)
 #endif
     if (!ipFlag->Quiet)
     {
+      ipFlag->error = errno;
       perror(_("dos2unix: Failed to open temporary output file"));
-      ipFlag->error = 1;
     }
     RetVal = -1;
   }
@@ -1097,8 +1097,8 @@ int ConvertDosToUnixOldFile(char* ipInFN, CFlag *ipFlag)
     }
   }
 
-  /* any error? */
-  if (RetVal)
+  /* any error? cleanup the temp file */
+  if (RetVal && (TempPath != NULL))
   {
     if (unlink(TempPath))
     {
@@ -1347,23 +1347,19 @@ int main (int argc, char *argv[])
             pFlag->FromToMode = FROMTO_MAC2UNIX;
           else
           {
-            if (!pFlag->Quiet)
-          {
-              fprintf(stderr, _("dos2unix: invalid %s conversion mode specified\n"),argv[ArgIdx]);
-              pFlag->error = 1;
-        }
+            fprintf(stderr, _("dos2unix: invalid %s conversion mode specified\n"),argv[ArgIdx]);
+            pFlag->error = 1;
             ShouldExit = 1;
+            pFlag->stdio_mode = 0;
           }
         }
         else
         {
           ArgIdx--;
-          if (!pFlag->Quiet)
-          {
-            fprintf(stderr,_("dos2unix: option '%s' requires an argument\n"),argv[ArgIdx]);
-            pFlag->error = 1;
-          }
+          fprintf(stderr,_("dos2unix: option '%s' requires an argument\n"),argv[ArgIdx]);
+          pFlag->error = 1;
           ShouldExit = 1;
+          pFlag->stdio_mode = 0;
         }
       }
 
@@ -1372,12 +1368,10 @@ int main (int argc, char *argv[])
         /* last convert not paired */
         if (!CanSwitchFileMode)
         {
-          if (!pFlag->Quiet)
-          {
-            fprintf(stderr, _("dos2unix: target of file %s not specified in new file mode\n"), argv[ArgIdx-1]);
-            pFlag->error = 1;
-          }
+          fprintf(stderr, _("dos2unix: target of file %s not specified in new file mode\n"), argv[ArgIdx-1]);
+          pFlag->error = 1;
           ShouldExit = 1;
+          pFlag->stdio_mode = 0;
         }
         pFlag->NewFile = 0;
       }
@@ -1387,20 +1381,18 @@ int main (int argc, char *argv[])
         /* last convert not paired */
         if (!CanSwitchFileMode)
         {
-          if (!pFlag->Quiet)
-          {
-            fprintf(stderr, _("dos2unix: target of file %s not specified in new file mode\n"), argv[ArgIdx-1]);
-            pFlag->error = 1;
-          }
+          fprintf(stderr, _("dos2unix: target of file %s not specified in new file mode\n"), argv[ArgIdx-1]);
+          pFlag->error = 1;
           ShouldExit = 1;
+          pFlag->stdio_mode = 0;
         }
         pFlag->NewFile = 1;
       }
       else { /* wrong option */
         PrintUsage();
         ShouldExit = 1;
-        if (!pFlag->Quiet)
-          pFlag->error = 1;
+        pFlag->error = 1;
+        pFlag->stdio_mode = 0;
       }
     }
     else
@@ -1493,7 +1485,7 @@ int main (int argc, char *argv[])
   }
 
 
-  if ((!pFlag->Quiet) && (!CanSwitchFileMode))
+  if (!CanSwitchFileMode)
   {
     fprintf(stderr, _("dos2unix: target of file %s not specified in new file mode\n"), argv[ArgIdx-1]);
     pFlag->error = 1;
